fix: Big Int type conversion for kotlin

+++ packages/kotlin-generator/src/helpers.ts
@@ -39,7 +39,7 @@ export function generateKotlinTypeName(type: Type): string {
       return "Double";
 
     case BigIntPrimitiveType:
-      return "BigInt";
+      return "BigInteger";
 
     case DatePrimitiveType:
     case DateTimePrimitiveType:
